Propagate tags to intermediate targets
The top-level "kustomized_resources" macro wraps a rule that accept a
"tags" attribute, and the rule generates a few intermediate targets to
collect runfiles and feed them to a launcher program that eventually
invokes the "kustomize" tool. If we don't propagate tags such as
"manual" that are present on the top-level targets to these
intermediate targets, then the intermediate targets may be included
unintentionally in wildcard target patterns, causing failures when
those targets' dependencies require a particular configuration to be
active.

In order to ensure that these intermediate targets behave equivalently
to the targets that motivated them, propagate the same tags specified
for the latter to the former.